inject logger again
